requirements/requirements.txt
basicsr>=1.4.2
facexlib>=0.2.0
gfpgan>=1.3.8
numpy==1.26.4
opencv-python>=4.5.5
pillow>=9.0.0
torch==2.0.1
torchvision==0.15.2
tqdm>=4.64.0
scipy>=1.7.3
lmdb>=1.3.0
pyyaml>=6.0
